config

+++ plugins/kanban/custom/updates/migration1019.php
@@ -7,19 +7,23 @@ class Migration1019 extends Migration
 {
     public function up()
     {
-        Schema::table('users', function($table)
-        {
-            $table->integer('team_id')->after('id');
-            $table->string('picture')->after('name')->nullable();
-        });
+        if (Schema::hasTable('users')) {
+            Schema::table('users', function($table)
+            {
+                $table->integer('team_id')->after('id');
+                $table->string('picture')->after('name')->nullable();
+            });
+        }
     }
 
     public function down()
     {
-        Schema::table('users', function($table)
-        {
-            $table->dropColumn('team_id');
-            $table->dropColumn('picture');
-        });
+        if (Schema::hasTable('users')) {
+            Schema::table('users', function($table)
+            {
+                $table->dropColumn('team_id');
+                $table->dropColumn('picture');
+            });
+        }
     }
 }
